Split VS Code serverReadyAction plumbing
Split the old launchProfiles module into dotnet launch settings, dotnet launch profile helpers, and an explicit VS Code serverReadyAction module so editor-specific behavior is organized as such.
